+++ src/components/mypage/MUserProfile.tsx
@@ -22,14 +22,14 @@ const USER_ROLE: { [id: string]: string } = {
 
 function MUserProfile({ userProfile }: UserProfileProps) {
     const { isModalOpen, openModal, closeModal } = useModal();
-    const [modalProps, setModalProps] = useState<ImodalProps>({
-        userid: -1,
+    const [modalProps, setModalProps] = useState<ImodalProps>(() => ({
+        userid: userProfile.id,
         follow: '',
-    });
+        isMine: userProfile.isMine,
+    }));
     const handleModal = (e: React.MouseEvent<HTMLDivElement>) => {
         let follow = e.currentTarget.dataset.type;
-        let userid = userProfile.id;
-        setModalProps({ userid, follow });
+        setModalProps({ ...modalProps, follow });
         openModal();
     };
